Refactored WebDriver to json_wire_spec.
